DEV: skip telescope migrations in production

+++ database/migrations/2024_11_19_114921_create_telescope_entries_table.php
@@ -18,6 +18,10 @@ return new class() extends Migration {
      */
     public function up(): void
     {
+        if (app()->isProduction()) {
+            return;
+        }
+
         $schema = Schema::connection($this->getConnection());
 
         $schema->create('telescope_entries', function (Blueprint $table) {
@@ -60,6 +64,10 @@ return new class() extends Migration {
      */
     public function down(): void
     {
+        if (app()->isProduction()) {
+            return;
+        }
+
         $schema = Schema::connection($this->getConnection());
 
         $schema->dropIfExists('telescope_entries_tags');
